// src/app/models/product.ts
export interface Product {
    itemNo: number;
    numPerPack: number;
    orderNo: number;
    packs: number;
    partNo: string;
    quantity: number;
    returnsNo: number;
    price: number;
    vat: number;
    vatRate: number;
    discrepancies: number;
    discount: number;
    creditNotes: number;
}